upload order api

+++ backend/api/models/cart.js
@@ -38,7 +38,7 @@ Cart.init(
       allowNull: false,
     },
     totalPrice: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
     },
     createdAt: {
